move list e2e tests to stop template subtest
